extracted codestream_local code
